Initial attempt at adding support for 'extra' arguments
